common: some clean up and warning squashing

See merge request graphviz/graphviz!3086

+++ lib/common/types.h
@@ -19,6 +19,7 @@
 #define WITH_CGRAPH 1
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdio.h>
 #include <assert.h>
 #include <signal.h>
@@ -250,12 +251,12 @@ extern "C" {
 
     typedef struct nlist_t {
 	node_t **list;
-	int size;
+	size_t size;
     } nlist_t;
 
     typedef struct elist {
 	edge_t **list;
-	int size;
+	size_t size;
     } elist;
 
 #define GUI_STATE_ACTIVE    (1<<0)
@@ -316,7 +317,7 @@ typedef enum {NATIVEFONTS,PSFONTS,SVGFONTS} fontname_kind;
 	/* connected components */
 	node_t *minset; /* set leaders */
 	node_t *maxset;	/* set leaders */
-	long n_nodes;
+	size_t n_nodes;
 	/* includes virtual */
 	int minrank;
 	int maxrank;
